separate main content for using personal center
